refactor: split settings pane metadata from the pane components
Settings/panes.ts holds the name, label, title, icon, group and rank of each
settings page and imports no components. Settings/index.ts attaches the
components and registers them.

The command palette lists the settings pages. It can now read the metadata
without pulling the panes, which build to a 1.1 MB chunk, into its own.

Part of extensibility phase 1, step 10.
