+++ frontend/src/components/Settings/index.ts
@@ -11,157 +11,38 @@ import PageCode from "@/components/Settings/PageCode.vue";
 import PageGeneral from "@/components/Settings/PageGeneral.vue";
 import PageMeta from "@/components/Settings/PageMeta.vue";
 import PageRobots from "@/components/Settings/PageRobots.vue";
-import { createRegistry, type RegistryItem } from "@/utils/createRegistry";
+import { settingsPanes, type SettingsPane } from "@/components/Settings/panes";
+import { createRegistry } from "@/utils/createRegistry";
 import type { Component } from "vue";
 
-export type SettingsGroup = "Current Page" | "Global";
+export { settingsGroups } from "@/components/Settings/panes";
+export type { SettingsGroup, SettingsPane } from "@/components/Settings/panes";
 
-export type SettingsItem = RegistryItem & {
-	label: string;
-	title: string;
-	icon: string;
-	group: SettingsGroup;
+export type SettingsItem = SettingsPane & {
 	component: Component;
-	disabled?: boolean;
 };
 
-// the sidebar renders groups in this order
-export const settingsGroups: SettingsGroup[] = ["Current Page", "Global"];
+const paneComponents: Record<string, Component> = {
+	page_general: PageGeneral,
+	page_code: PageCode,
+	page_meta: PageMeta,
+	page_analytics: PageAnalytics,
+	global_general: GlobalGeneral,
+	global_users: GlobalUsers,
+	global_code: GlobalCode,
+	global_redirects: GlobalRedirects,
+	global_robots: PageRobots,
+	global_domains: GlobalDomains,
+	global_analytics: GlobalAnalytics,
+	global_developer: GlobalDeveloper,
+	global_ai: GlobalAI,
+};
 
 export const settingsItems = createRegistry<SettingsItem>();
 export const registerSettingsItem = settingsItems.register;
 
-// name doubles as the persisted settingsActiveTab value, so it must not change
 export function registerBuiltInSettingsItems() {
-	registerSettingsItem({
-		name: "page_general",
-		label: "General",
-		title: "General",
-		icon: "lucide-settings",
-		group: "Current Page",
-		rank: 10,
-		component: PageGeneral,
-	});
-
-	registerSettingsItem({
-		name: "page_code",
-		label: "Code",
-		title: "Page Code",
-		icon: "lucide-code",
-		group: "Current Page",
-		rank: 20,
-		component: PageCode,
-	});
-
-	registerSettingsItem({
-		name: "page_meta",
-		label: "Meta",
-		title: "Meta",
-		icon: "lucide-square-dashed-bottom-code",
-		group: "Current Page",
-		rank: 30,
-		component: PageMeta,
-	});
-
-	registerSettingsItem({
-		name: "page_analytics",
-		label: "Analytics",
-		title: "Page Analytics",
-		icon: "lucide-chart-bar",
-		group: "Current Page",
-		rank: 40,
-		component: PageAnalytics,
-	});
-
-	registerSettingsItem({
-		name: "global_general",
-		label: "General",
-		title: "General",
-		icon: "lucide-settings",
-		group: "Global",
-		rank: 110,
-		component: GlobalGeneral,
-		disabled: false,
-	});
-
-	registerSettingsItem({
-		name: "global_users",
-		label: "Users",
-		title: "Users",
-		icon: "lucide-users",
-		group: "Global",
-		rank: 120,
-		component: GlobalUsers,
-	});
-
-	registerSettingsItem({
-		name: "global_code",
-		label: "Code",
-		title: "Global Code",
-		icon: "lucide-code",
-		group: "Global",
-		rank: 130,
-		component: GlobalCode,
-	});
-
-	registerSettingsItem({
-		name: "global_redirects",
-		label: "Redirects",
-		title: "Redirects",
-		icon: "lucide-shuffle",
-		group: "Global",
-		rank: 140,
-		component: GlobalRedirects,
-	});
-
-	registerSettingsItem({
-		name: "global_robots",
-		label: "Robots",
-		title: "Robots.txt",
-		icon: "lucide-bot",
-		group: "Global",
-		rank: 150,
-		component: PageRobots,
-	});
-
-	registerSettingsItem({
-		name: "global_domains",
-		label: "Domains",
-		title: "Custom Domains",
-		icon: "lucide-globe",
-		group: "Global",
-		rank: 160,
-		component: GlobalDomains,
-		condition: () => Boolean(window.is_fc_site || window.is_developer_mode),
-	});
-
-	registerSettingsItem({
-		name: "global_analytics",
-		label: "Analytics",
-		title: "Site Analytics",
-		icon: "lucide-chart-bar",
-		group: "Global",
-		rank: 170,
-		component: GlobalAnalytics,
-	});
-
-	registerSettingsItem({
-		name: "global_developer",
-		label: "Developer",
-		title: "Developer Settings",
-		icon: "lucide-terminal",
-		group: "Global",
-		rank: 180,
-		component: GlobalDeveloper,
-	});
-
-	registerSettingsItem({
-		name: "global_ai",
-		label: "AI",
-		title: "AI Settings",
-		icon: "lucide-sparkles",
-		group: "Global",
-		rank: 190,
-		component: GlobalAI,
-	});
+	settingsPanes.forEach((pane) =>
+		registerSettingsItem({ ...pane, component: paneComponents[pane.name] }),
+	);
 }
